pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "measurement-experimentation"
version = "0.1.0"
description = "Agentic framework for eliciting and designing ad campaign measurement experiments"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain-anthropic>=0.3",
    "langchain-core>=0.3",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "streamlit>=1.41",
    "pydantic>=2.9",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "pymc>=5.18",
    "numpy>=1.26",
    "pandas>=2.2",
    "markdown2>=2.4",
    "jinja2>=3.1",
    "arviz>=0.20",
    "scipy>=1.12",
    "plotly>=5.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["measurement_design*", "backend*", "frontend*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
